+++ src/client.ts
@@ -28,7 +28,7 @@ export interface Client {
    * By default it performs **GET** request to _path_ URL.
    */
   request: (
-    path: string,
+    path: string | URL,
     options?: ClientSessionRequestOptions
   ) => Promise<Response>;
   /** Closes connection with remote peer. */
@@ -63,7 +63,7 @@ const createClient = (
       const stream: ClientHttp2Stream = instance
         .request(
           {
-            [constants.HTTP2_HEADER_PATH]: path,
+            [constants.HTTP2_HEADER_PATH]: path.toString(),
             [constants.HTTP2_HEADER_METHOD]: constants.HTTP2_METHOD_GET,
             ...headers,
           },